Move StrahlkorperDiffFrame to Domain
StrahlkorperInDifferentFrame and StrahlkorperCoordsInDifferentFrame
both depend on Domain, so it made sense to put it there.
